Move feeds store to module
