+++ packages/client-web/src/store/index.ts
@@ -4,7 +4,7 @@ import { composeWithDevTools } from "redux-devtools-extension";
 
 import user from "./user";
 import date from "../modules/date/model/store";
-import feeds from "./feeds";
+import feeds from "../modules/feeds/model/store";
 import timeline from "../modules/timeline/model/store";
 
 const reducer = combineReducers({
